Change personalNumber to Long and length to DOuble

+++ src/Model/Log.java
@@ -97,7 +97,7 @@ public class Log {
 
 
 
-    public void changeMem(String name , int num,String id){
+    public void changeMem(String name , long num,String id){
 
         BufferedWriter writer = null;
         String mapID = "Files\\"+id;
@@ -118,7 +118,7 @@ public class Log {
 
     }
 
-    public void changeBoat(Boat boat, String memID,int newLength, String newModel){
+    public void changeBoat(Boat boat, String memID,double newLength, String newModel){
 
         BufferedWriter writer = null;
         int boatID = boat.nr;
